Move utility helper code in a separate module
