+++ src/utils/mod.rs
@@ -1,0 +1,7 @@
+// ---------
+// UTILITIES
+// ---------
+
+pub mod ansi;
+pub mod format;
+pub mod helpers;
